Move entrypoint logic to separate module
