+++ gaphor/application.py
@@ -8,20 +8,22 @@ can be edited.
 
 from __future__ import annotations
 
-import inspect
 import logging
-from typing import Dict, Iterator, Optional, Set, Tuple, Type, TypeVar
+from typing import Dict, Iterator, Optional, Set, Tuple, Type, TypeVar, cast
 
 import importlib_metadata
 
 from gaphor import transaction
 from gaphor.abc import ActionProvider, Service
 from gaphor.action import action
+from gaphor.core.eventmanager import EventManager
+from gaphor.entrypoint import init_entrypoints, load_entrypoints
 from gaphor.event import (
     ServiceInitializedEvent,
     ServiceShutdownEvent,
     SessionShutdownRequested,
 )
+from gaphor.services.componentregistry import ComponentRegistry
 
 T = TypeVar("T")
 
@@ -40,10 +42,6 @@ class NotInitializedError(Exception):
     pass
 
 
-class ComponentLookupError(LookupError):
-    pass
-
-
 class Application(Service, ActionProvider):
     """
     The Gaphor application is started from the gaphor.ui module.
@@ -55,12 +53,16 @@ class Application(Service, ActionProvider):
     are registered in the "component_registry" service.
     """
 
-    def __init__(self, appservices=None):
+    def __init__(self):
         self.active_session: Optional[Session] = None
         self.sessions: Set[Session] = set()
 
-        uninitialized_services = load_services("gaphor.appservices", appservices)
-        self._services_by_name = init_services(uninitialized_services, application=self)
+        uninitialized_services: Dict[str, Type[Service]] = load_entrypoints(
+            "gaphor.appservices"
+        )
+        self._services_by_name = init_entrypoints(
+            uninitialized_services, application=self
+        )
 
         transaction.subscribers.add(self._transaction_proxy)
 
@@ -134,11 +136,17 @@ class Session:
         """
         Initialize the application.
         """
-        uninitialized_services = load_services("gaphor.services", services)
-        services_by_name = init_services(uninitialized_services)
+        uninitialized_services: Dict[str, Type[Service]] = load_entrypoints(
+            "gaphor.services", services
+        )
+        services_by_name = init_entrypoints(uninitialized_services)
 
-        self.event_manager = services_by_name["event_manager"]
-        self.component_registry = services_by_name["component_registry"]
+        self.event_manager: EventManager = cast(
+            EventManager, services_by_name["event_manager"]
+        )
+        self.component_registry: ComponentRegistry = cast(
+            ComponentRegistry, services_by_name["component_registry"]
+        )
 
         for name, srv in services_by_name.items():
             logger.info(f"Initializing service {name}")
@@ -153,76 +161,12 @@ class Session:
 
     def shutdown(self):
         if self.component_registry:
-            for name, _srv in self.component_registry.all(Service):
-                self.shutdown_service(name)
+            for name, srv in self.component_registry.all(Service):  # type: ignore[misc]
+                self.shutdown_service(name, srv)
 
-        self.component_registry = None
-        self.event_manager = None
-
-    def shutdown_service(self, name):
+    def shutdown_service(self, name, srv):
         logger.info(f"Shutting down service {name}")
-        assert self.component_registry
-        assert self.event_manager
 
-        srv = self.component_registry.get_service(name)
         self.event_manager.handle(ServiceShutdownEvent(name, srv))
         self.component_registry.unregister(srv)
         srv.shutdown()
-
-
-def load_services(scope, services=None) -> Dict[str, Type[Service]]:
-    """
-    Load services from resources.
-
-    Service should provide an interface `gaphor.abc.Service`.
-    """
-    uninitialized_services = {}
-    for ep in importlib_metadata.entry_points()[scope]:
-        cls = ep.load()
-        if isinstance(cls, Service):
-            raise NameError(f"Entry point {ep.name} doesnt provide Service")
-        if not services or ep.name in services:
-            logger.debug(f'found service entry point "{ep.name}"')
-            uninitialized_services[ep.name] = cls
-    return uninitialized_services
-
-
-def init_services(uninitialized_services, **known_services):
-    """
-    Instantiate service definitions, taking into account dependencies
-    defined in the constructor.
-
-    Given a dictionary `{name: service-class}`,
-    return a map `{name: service-instance}`.
-    """
-    ready: Dict[str, Service] = dict(known_services)
-
-    def pop(name):
-        try:
-            return uninitialized_services.pop(name)
-        except KeyError:
-            return None
-
-    def init(name, cls):
-        kwargs = {}
-        for dep in inspect.signature(cls).parameters:
-            if dep not in ready:
-                depcls = pop(dep)
-                if depcls:
-                    kwargs[dep] = init(dep, depcls)
-                else:
-                    logger.info(
-                        f"Service {name} parameter {dep} does not reference a service"
-                    )
-            else:
-                kwargs[dep] = ready[dep]
-        srv = cls(**kwargs)
-        ready[name] = srv
-        return srv
-
-    while uninitialized_services:
-        name = next(iter(uninitialized_services.keys()))
-        cls = pop(name)
-        init(name, cls)
-
-    return ready
